add docappender.blocked.add counter
This monotonic counter increments every time Appender.Add
detects that bulkItems channel length and capacity are
equal before sending the created BulkIndexerItem to the
channel.

As bulkItems is a buffered channel, monitoring length and
capacity allow us to detect possible blocking events when
sending to the channel due to exhausted capacity.

As Appender.Add could be called from multiple goroutines,
we are not guaranteed that every time this metric is
incremented the channel send will block, but should give us
a reasonably approximated picture.